requirements/dev.txt
-r base.txt
ipdb
django-extensions
django-devserver